Complete redesign: Professional UI, CRUD operations, Flatpickr calendar, mobile responsive, deployment ready

+++ app1/models.py
@@ -6,4 +6,4 @@ class Student(models.Model):
     age=models.IntegerField()
     email=models.EmailField(max_length=25)
     dob=models.DateField(max_length=20)
-    cell=models.IntegerField()
+    cell=models.BigIntegerField()
